requirements.txt
asgiref==3.10.0
channels==4.3.1
channels_redis==4.3.0
Django==5.2.8
djangorestframework==3.16.1
msgpack==1.1.2
python-dotenv==1.2.1
redis==7.0.1
sqlparse==0.5.3
tzdata==2025.2
